pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "riva-dictation"
version = "1.0.0"
description = "Real-time speech-to-text with NVIDIA Riva"
readme = "README.md"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "nvidia-riva-client",
    "pyaudio",
    "numpy",
    "pyautogui",
    "pynput",
    "infi.systray",
    "pillow",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0"
]

[project.scripts]
riva-dictation = "riva_dictation.cli:main"

[tool.setuptools]
packages = ["riva_dictation"]

[tool.pytest.ini_options]
testpaths = ["riva_dictation/tests"]
python_files = ["test_*.py"]
addopts = "--cov=riva_dictation --cov-report=term-missing"

[tool.coverage.run]
source = ["riva_dictation"]
omit = ["riva_dictation/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
    "raise ImportError"
]